refactor Config 클래스 common으로 이동
- Config 클래스별 관리의 편의를 위한 디렉토리 이동
