+++ webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/common/OpenSearchClientConfig.java
@@ -1,4 +1,4 @@
-package com.samsamohoh.webtoonsearch.adapter.persistence.searchengine.config;
+package com.samsamohoh.webtoonsearch.common;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.http.HttpHost;
